+++ app/models/user.rb
@@ -4,6 +4,7 @@ class User < ApplicationRecord
   has_many :transactions, dependent: :destroy
   has_many :accounts, dependent: :destroy
   has_many :categories, dependent: :destroy
+  has_many :budgets, dependent: :destroy
 
   normalizes :email_address, with: ->(e) { e.strip.downcase }