+++ config/routes.rb
@@ -1,6 +1,8 @@
 Rails.application.routes.draw do
   mount RailsIcons::Engine, at: "/rails_icons"
   get "home/index"
+  get "dashboard" => "dashboard#show", as: :dashboard
+  get "reports" => "reports#show", as: :reports
   resource :session
   resources :passwords, param: :token
   resource :settings, only: %i[ edit update ]
@@ -12,6 +14,7 @@ Rails.application.routes.draw do
   resources :accounts, except: :show
   resources :categories, except: :show
   resources :transactions, except: :show
+  resources :budgets, except: :show
   resources :transfers, only: %i[ new create edit update destroy ]
 
   namespace :admin do